v4l2, fmr: Increase the range to allow to read all the FMR IP registers
In the current solution, only a limited range of registers are allowed
to be read by the application. Changes are made to allow to read all the
registers of FM Radio IP.

Check patch errors are also corrected.

UTP-Id: SMS05606433

+++ drivers/media/radio/intel/xgold632/FMR_HLD/aud_app_fmr_hld_rx.h
@@ -313,7 +313,7 @@ struct fmrx_state {
 	enum fmrx_rds_pi_mode rds_pi_mode_bkp;
 	enum fmtrx_subscribe_rds_mode rds_mode_bkp;
 
-	s32 af_eval_state; /* State of AF evaluation */
+	u8 af_eval_state; /* State of AF evaluation */
 	s32 idi_hs_ctrl;   /* State of FMR IDI handshake control bit */
 
 	/* Results from last channel evaluation */
